backend/frontend/requirements.txt
# Core dependencies
streamlit==1.29.0
pandas==2.1.3
numpy==1.24.3
plotly==5.18.0
requests==2.31.0

# UI components
streamlit-option-menu==0.3.6
streamlit-authenticator==0.3.2
streamlit-extras==0.3.5

# Data processing
openpyxl==3.1.2
xlrd==2.0.1

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
wordcloud==1.9.2

# Utilities
python-dotenv==1.0.0
pillow==10.1.0
pytz==2023.3

# WebSocket
websocket-client==1.7.0

# File handling
python-multipart==0.0.6

# Date handling
python-dateutil==2.8.2